Move state to a separated file
